requirements.txt
certifi==2023.07.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-default-group-wheel==1.2.2
cryptography==41.0.3
Deprecated==1.2.13
idna==3.4
prettytable==3.8.0
pycparser==2.21
pygit2==1.12.2
PyJWT==2.7.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlite-fts4==1.0.3
sqlite-utils==3.30
tabulate==0.9.0
urllib3==1.26.15
wcwidth==0.2.6
wrapt==1.15.0
